OneDrive/Desktop/Hire_AI/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
supabase==2.3.1
pypdf2==3.0.1
openai==1.12.0
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 